chore: move lore data around
