+++ src/lore-data/archetypes.json
