+++ src/lore-data/story.json
